requirements.txt
# Google Signals Bot Requirements

# Основные зависимости
python-dotenv==1.0.0
requests==2.31.0

# Bybit API
pybit==5.7.0

# Google Sheets API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Логирование и утилиты
colorama==0.4.6 